+++ FlagsRally_xunit/FlagsRally_xunit/SubRegionHelperTest.cs
@@ -1,5 +1,4 @@
 ﻿using FlagsRally.Helpers;
-using FlagsRally.Utilities;
 
 
 namespace FlagsRallyTests.Helpers;
